Sort controller
