+++ src/controllers/trip.js
@@ -1,54 +1,44 @@
 import moment from "moment/moment";
 
 import Point from "./point";
-import Sorter from '../components/sorter';
 import DaysList from '../components/days-list';
-
+import Sort from "./sort";
 import Day from '../components/day';
-import {hideVisually, Position, render, showVisually} from '../utils';
-import {calculateEventCost, eventTypes, EventCategories} from "../data";
+
 import {Mode} from "../components/event-editor";
+import {hideVisually, showVisually, Position, render} from '../utils';
+import {eventTypes, EventCategories} from "../data";
+
 
 export default class Trip {
   constructor(container, events) {
     this._container = container;
     this._events = events;
-
-    this._sorter = new Sorter();
+    this._sorter = new Sort(this._container, this._events, this._onSortChanged.bind(this));
     this._dayList = new DaysList();
 
     this._subscriptions = [];
     this._creatingEvent = null;
   }
 
-  show() {
-    showVisually(this._container);
-  }
-
-  hide() {
-    hideVisually(this._container);
-  }
-
   init() {
     if (this._events.length) {
       // Sorter
-      render(this._container, this._sorter.getElement(), Position.BEFOREEND);
+      this._sorter.renderSort();
 
-      // DAY LIST
+      // Day list
       render(this._container, this._dayList.getElement(), Position.BEFOREEND);
     }
 
-    this._renderEvents();
-
-    this._sorter.getElement()
-      .addEventListener(`click`, (evt) => this._onSortLinkClick(evt));
+    const sortedEvents = this._sorter.sort(this._events);
+    this._renderEvents(sortedEvents);
   }
 
   /**
    * @private
    */
-  _renderEvents() {
-    const groupedDays = this._sort();
+  _renderEvents(groupedDays) {
+    this._dayList.getElement().innerHTML = ``;
     const allDays = groupedDays.map(({date, events}, index) => new Day({day: date, number: (index + 1), events}));
 
     if (allDays.length) {
@@ -66,110 +56,6 @@ export default class Trip {
       const pointController = new Point(this._container, null, Mode.CREATING, this._onDataChange.bind(this), this._onViewChange.bind(this));
       this._subscriptions.push(pointController.setDefaultView.bind(pointController));
     }
-  }
-
-  /**
-   *
-   * @param evt
-   * @private
-   */
-  _onSortLinkClick(evt) {
-    if (evt.target.tagName !== `LABEL`) {
-      return;
-    }
-
-    this._sort(evt.target.dataset.sort);
-  }
-
-  /**
-   * Сортировка точек в соответствии с типом сортировки
-   *
-   * @param sortType
-   * @private
-   */
-  _sort(sortType = `sort-default`) {
-    this._dayList.getElement().innerHTML = ``;
-    let days = [];
-
-    switch (sortType) {
-      case `sort-price`:
-        days = this._sortByPrice(this._events);
-        break;
-      case `sort-time`:
-        days = this._sortByTime(this._events);
-        break;
-      case `sort-default`:
-        days = this._defaultSort(this._events);
-        break;
-    }
-
-    return days;
-  }
-
-  /**
-   * Сортировка по умолчанию:
-   *
-   * Группировка по дням в хронологическом порядке,
-   * день события, это день в котором оно начинается
-   *
-   * @param {array} events
-   * @returns {{date: string, events:*}[]}
-   * @private
-   */
-  _defaultSort(events) {
-    const groupedEvents = {};
-
-    events.forEach((event) => {
-      const dateString = event.from.toDateString();
-
-      if (!Array.isArray(groupedEvents[dateString])) {
-        groupedEvents[dateString] = [];
-      }
-
-      groupedEvents[dateString].push(event);
-    });
-
-    let groupedDays = Object.keys(groupedEvents).map((date) => {
-      const events = groupedEvents[date];
-
-      return {date, events};
-    });
-
-    groupedDays.sort((dayA, dayB) => ((new Date(dayA.date)).getTime() - (new Date(dayB.date)).getTime()));
-
-    return groupedDays;
-  }
-
-  /**
-   * Сортировка по цене
-   *
-   * @param {array} events
-   * @returns {{date: null, events: *}[]}
-   * @private
-   */
-  _sortByPrice(events) {
-    events.sort((eventA, eventB) => calculateEventCost(eventB) - calculateEventCost(eventA));
-
-    return [{
-      date: null,
-      events
-    }];
-  }
-
-  /**
-   * Сортировка по длительности события
-   *
-   * @param {array} events
-   * @returns {{date: null, events: *}[]}
-   * @private
-   */
-  _sortByTime(events) {
-    events.sort((eventA, eventB) => (eventB.to - eventB.from) - (eventA.to - eventA.from));
-
-    return [{
-      date: null,
-      events
-    }];
   }
 
   _onDataChange(newData, id) {
@@ -190,11 +76,16 @@ export default class Trip {
       this._events[index] = newData;
     }
 
-    this._renderEvents();
+    const sortedEvents = this._sorter.sort(this._events);
+    this._renderEvents(sortedEvents);
   }
 
   _onViewChange() {
     this._subscriptions.forEach((subscription) => subscription());
+  }
+
+  _onSortChanged(sortedEvents) {
+    this._renderEvents(sortedEvents);
   }
 
   createEvent() {
@@ -223,5 +114,13 @@ export default class Trip {
         this._onDataChange.bind(this),
         this._onViewChange.bind(this)
       );
+  }
+
+  show() {
+    showVisually(this._container);
+  }
+
+  hide() {
+    hideVisually(this._container);
   }
 }
